Carry the advertised friendly name in the server-info handshake
