+++ esphome_device_builder/models/api.py
@@ -108,6 +108,9 @@ class ServerInfoMessage(DashboardModel):
     esphome_version: str
     port: int
     ha_addon: bool = False
+    # Advertised display name (mDNS TXT / pairing handshake); the frontend
+    # maps the HA-addon container hostname to "Home Assistant App".
+    friendly_name: str = ""
     # True only when the connection is proxied through HA ingress; an
     # add-on reached directly on its exposed port is False, unlike ha_addon.
     ha_ingress: bool = False
